Issue #21 - Move appsettings to separate file
